kv: don't clobber commit status when heartbeat fails
When a heartbeat failed while a request was in-flight, two proto updates
could race and produce illegal behavior. For example, it was possible
for a client to successfully commit but receive an ABORTED proto (and no
error).

This commit removes the ad-hoc manipulation of the state during
heartbeat failures. Instead, before each new request sent through the
heartbeat interceptor we check on the heartbeat loop's status; if it
indicates an error, we refuse new requests.

Fixes #39658.

Release note: None